requirements.txt
django==4.2.26
djangorestframework==3.14.0
flake8>=7.0.0
drf-spectacular==0.29.0
